utilisation de le defense + visualisation des PaidChest + graphisme #57 #85

+++ RogueLikeMADY/src/main/java/com/mady/utils/entities/Amulet.java
@@ -9,11 +9,11 @@ public class Amulet extends AbstractStuffItem {
      * @param AGI double
      * @param LUK double
      */
-    public Amulet(double HP, double MP, double ATK, double DEF, double AGI, double LUK) {
+    public Amulet(int HP, int MP, int ATK, int DEF, int AGI, int LUK) {
         super("amulet", HP, MP, ATK, DEF, AGI, LUK);
     }
 
-    public Amulet(double LUK) {
+    public Amulet(int LUK) {
         this(0, 0, 0, 0, 0, LUK);
     }
